// projects/facility-management/src/app/dashboard/dashboard.enum.ts
/** This is a enum  for ChartColor which contains color properties */
export enum ChartColor {
    Blue = '#22a8ee',
    Yellow = '#ffbf3a',
    Orange = '#ff8f2b',
    Purple = '#763fea',
    Red = '#fa3a4b',
    Black = '#000000',
    Brown = '#795548',
    Gray = '#9EA2AE',
    Green = '#2ec740',
}
/** This is a enum  for ComboChartParam which contains month values */
export enum ComboChartParam {
    Monthly = 1,
    Quarterly = 2,
    Annually = 3,
}

/** This is a enum  for ComboChartLabel  */
export enum ComboChartLabel {
    Monthly = 'Monthly',
    Quarterly = 'Quarterly',
    Annually = 'Annual',
}
/** This is a enum  for OpenRequestTitle */
export enum OPEN_REQUEST_TITLE {
    Copyit = 1,
    Fleet = 2,
    Bookit = 3,
}

/** This is a enum  for OpenRequestLabel */
export enum OPEN_REQUEST_LABEL {
    Copyit = 'Copy It',
    Bookit = 'Book It',
    Fleet = 'Fleet Management',
}

/** This is a enum to set the line for bar */
export enum BAR_LINE {
    Line = 4
}

export enum BARS_PER_WINDOW {
    Bar_Max = 6,
    Bar_Min = 0
}